requirements.txt
fastapi
pytest
requests
uvicorn
allure-pytest
allure-python-commons